requirements.txt
numpy==1.13.0
opencv-python==3.2.0.7
PyYAML==3.12
